+++ docs/meta/history.md
@@ -125,6 +125,10 @@ This is the first release candidate for Hatch v1, a complete rewrite.
 
 - In order to simplify configuration, metadata hooks are now configured under `tool.hatch.metadata.hooks` rather than directly under `tool.hatch.metadata`
 
+***Added:***
+
+- Add option to allow the use of direct references for dependencies
+
 ### [0.20.1](https://github.com/ofek/hatch/releases/tag/hatchling-v0.20.1) - 2022-03-07 ### {: #hatchling-v0.20.1 }
 
 ***Fixed:***